+++ src/dom/NuBuilder.ts
@@ -8,7 +8,7 @@ import type {
     DOMEventHandlers
 } from "../types.ts";
 import { extend } from "./nu.ts";
-
+import { CfDom } from "./config.ts";
 
 /**
  * Creates a typed HTML element based on the tag name.
@@ -18,7 +18,7 @@ import { extend } from "./nu.ts";
  * @internal
  */
 const createTypedElement = <K extends keyof HTMLElementTagNameMap>(name: K) => {
-    return document.createElement(name);
+    return CfDom.createElement(name);
 }
 
 /**
@@ -96,7 +96,7 @@ export class NuBuilder<T extends string, E extends InferElementType<T>, D extend
         const elem = createTypedElement(tag as keyof HTMLElementTagNameMap);
 
         if (id) elem.id = id;
-        classes.forEach((cls) => elem.classList.add(cls));
+        if (classes?.length) classes.forEach(cls => elem.classList.add(cls));
 
         return extend(elem as E, this.props);
     }